Create password validator
Write a function CheckPassword(str) which will accept the string as an
argument or parameter and validates the password. It will return 1 if the
conditions are satisfied else it’ll return 0?
The password is valid if it satisfies the below conditions:
a. It should contain at least 4 characters.
b. At least 1 numeric digit should be present.
c. At least 1 Capital letter should be there.
d. Passwords should not contain space or slash(/).
e. The starting character should not be a number.
Sample Test Case:
Input:
bB1_89
Output:
1
Approach:
1. Using if condition check whether the length of string is greater than equal
to 4 or not.
2. Run a loop on the string and check if any character is a digit or not. It is a
digit if it’s between ‘0’ and ‘9’.
3. Iterate the string and check that only at least one letter should be between
‘A’ and ‘Z’, i.e it should be capital.
4. Run a loop on string and check no character should match space(‘ ‘) or
slash (‘/’).
5. Check that the first character should not lie between ‘0’ and ‘9’.