Create OperationChoices
Write a function OperationChoices(c, a, b) which will accept three integers
as an argument, and the function will return:
a. (a + b) if the value of c=1.
b. (a – b) if the value of c=2.
c. (a b) if the value of c=3.
d. (a / b) if the value of c=4.