Create sum of second largest in sub arrays
You are required to input the size of the matrix then the elements of matrix,
then you have to divide the main matrix in two sub matrices (even and odd)
in such a way that element at 0 index will be considered as even and
element at 1st index will be considered as odd and so on. Then you have sort
the even and odd matrices in ascending order and print the sum of second
largest number from both the matrices.