Create autobiographical number
An Autobiographical Number is a number N such that the first digit of N represents the count of how many zeroes are there in N, the second digit represents the count of how many ones are there in N and so on. You are given a function, def FindAutoCount(n): The function accepts string 'n which is a number and checks whether the number is an autobiographical number or not. If it is, an integer is returned, Le. the count of definct numbers in 'n'. If not, it returns 0.