Create equilibrium point
The function accepts an integer array ‘arr’ of size ‘n’ as its argument. The
function needs to return the index The function accepts an integer array ‘arr’ of size ‘n’ as its argument. The
function needs to return the index of an equilibrium point in the array, where
the sum of elements on the left of the index is equal to the sum of elements on
the right of the index. If no equilibrium point exists, the function should return
-1.The function accepts an integer array ‘arr’ of size ‘n’ as its argument. The
function needs to return the index of an equilibrium point in the array, where
the sum of elements on the left of the index is equal to the sum of elements on
the right of the index. If no equilibrium point exists, the function should return
-1.of an equilibrium point in the array, where
the sum of elements on the left of the index is equal to the sum of elements on
the right of the index. If no equilibrium point exists, the function should return
-1.